updated copyright, fixed some comments

+++ src/main/java/synapticloop/templar/function/comparison/FunctionGreaterThan.java
@@ -1,7 +1,7 @@
 package synapticloop.templar.function.comparison;
 
 /*
- * Copyright (c) 2012-2019 synapticloop.
+ * Copyright (c) 2012-2023 synapticloop.
  * All rights reserved.
  *
  * This source code and any derived binaries are covered by the terms and
@@ -26,7 +26,7 @@ public class FunctionGreaterThan extends FunctionNumericComparison {
 	protected Object evaluateFunction(Object[] args, TemplarContext templarContext) throws FunctionException {
 		super.evaluateNumeric(args, templarContext);
 
-		return(arg1.longValue() > arg2.longValue());
+		return(arg1.doubleValue() > arg2.doubleValue());
 	}
 
 }
